Schedule jobs from stat
Historically table audit_event was used to schedule jobs
Since stat table was introduced to show similar info, it is more convenient to use it for scheduling.
Events for miss and hit are not logged anymore.
New columns folder_id and file_id are added to stat table, they are used to distinguish path_miss vs mirror_miss.